Create button_pio.py
Hi,

I am not sure if this is the correct way to do this (propose this code), and I apologize in advance if it is not.  I tweaked the encoder example to create a button example.  This is my first time playing with PIO.  I'm pretty sure the read pin state loop could be only 2-3 lines long, but I left things as they were.  I felt this library needed a button example, for the next Parsko that shows up.

I'm unclear what line 34 is about.  I see the need I think, but don't understand what or why "set y 31".  Is it arbitrary?  

I plan on trying the encoder example with some higher line count rotaries to see if it loses counts.  So, I still need to understand the code where this came from.

Thanks,

-Parsko