Add `Program` class, type information
An instance of the `Program` class has the `.assembled` property
as well as the `pio_kwargs` property. This allows passing information
from the assembler to the StateMachine constructor, in a way that can
be extended in future versions (e.g., for .wrap / .wrap_target)

This re-indents a lot of code so it may be better viewed with whitespace
changes hidden.

Tests are added for the limits of the side-set and delay values. An
off-by-one bug with the sideset_value limit was fixed.